pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "thiagola92", email = "thiagola92@gmail.com"},
]
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn[standard]>=0.23.2",
    "SQLAlchemy>=2.0.20",
    "asyncpg>=0.28.0",
    "pydantic-settings>=2.0.3",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
start = "uvicorn src.main:app --reload"

